refactor: prompts
