+++ src/prompts/chart-prompts.ts
@@ -1,0 +1,20 @@
+export const chartPrompts = Object.freeze({
+  dailyOmzetHeatmap: '<EMPTY PROMPT>',
+  omzetJamHariHeatmap: '<EMPTY PROMPT>',
+  salesTrendHourlyDaily: '<EMPTY PROMPT>',
+  tcApcHarian: '<EMPTY PROMPT>',
+  omzetHarian: '<EMPTY PROMPT>',
+  omzetMingguan: '<EMPTY PROMPT>',
+  omzetOutlet: '<EMPTY PROMPT>',
+  penjualanBulanan: '<EMPTY PROMPT>',
+  penjualanChannel: '<EMPTY PROMPT>',
+  orderByCategory: '<EMPTY PROMPT>',
+  topMakanan: '<EMPTY PROMPT>',
+  topMinuman: '<EMPTY PROMPT>',
+  cabangOmzetCheck: '<EMPTY PROMPT>',
+  yoyOmzet: '<EMPTY PROMPT>',
+  cabangApc: '<EMPTY PROMPT>',
+  cabangDetail: '<EMPTY PROMPT>',
+  yoyDetail: '<EMPTY PROMPT>',
+  generalPdfInsights: '<EMPTY PROMPT>',
+});
