+++ src/prompts/index.ts
@@ -1,0 +1,3 @@
+export { viewPromptCreators } from './view-prompts';
+export { chartPrompts } from './chart-prompts';
+export const prompts = Object.freeze({});
